Producing plot "coverage - mean error" for applicability domain studies Prediction variance is taken as distance to model

Producing plot "coverage - mean error" for applicability domain studies Prediction variance is taken as distance to model

cmf_ecvr_plot_coverage_mean_error(ecvr_fname = "ligands-ecvr.RData", cmep_type = "in", smoothing = TRUE, ...)

Arguments

ecvr_fname

cmep_type

smoothing

…

Details

Value

References

Note

See also

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (ecvr_fname = "ligands-ecvr.RData", cmep_type = "in",
    smoothing = TRUE, ...)
{
    load(ecvr_fname)
    y_err <- abs(ecvr$y_exp - ecvr$y_pred_mean)
    d2m <- ecvr$y_pred_sd
    cinf_plot_coverage_mean_error(y_err, d2m, cmep_type = cmep_type,
        smoothing = smoothing, ...)
  }

#> function (ecvr_fname = "ligands-ecvr.RData", cmep_type = "in", 
#>     smoothing = TRUE, ...) 
#> {
#>     load(ecvr_fname)
#>     y_err <- abs(ecvr$y_exp - ecvr$y_pred_mean)
#>     d2m <- ecvr$y_pred_sd
#>     cinf_plot_coverage_mean_error(y_err, d2m, cmep_type = cmep_type, 
#>         smoothing = smoothing, ...)
#>   }
#> <environment: 0x10e515b18>