Computes CMF kernel matrices for prediction and saves them to file

Computes CMF kernel matrices for prediction and saves them to file

comp_kernels_pred(train_fname = "ligands-train.mol2", pred_fname = "ligands-pred.mol2", kernels_pred_fname = "ligands-kernels-pred.RData", mfields = c("q", "vdw", "logp", "abra", "abrb"), print_comp_kernels = TRUE, ...)

Arguments

train_fname

pred_fname

kernels_pred_fname

mfields

print_comp_kernels

…

Details

Value

References

Note

See also

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (train_fname = "ligands-train.mol2", pred_fname = "ligands-pred.mol2",
    kernels_pred_fname = "ligands-kernels-pred.RData", mfields = c("q",
        "vdw", "logp", "abra", "abrb"), print_comp_kernels = TRUE,
    ...)
{
    mdb0_train <- read_mol2(train_fname)
    mdb0_pred <- read_mol2(pred_fname)
    mdb_train <- cmf_params_tripos(mdb0_train)
    mdb_pred <- cmf_params_tripos(mdb0_pred)
    nfields <- length(mfields)
    syb_types <- get_syb_types_list(mdb_train)
    kernels_pred <- list()
    kernels_pred$alphas <- alphas
    for (f in 1:nfields) {
        kernels_pred[[mfields[f]]] <- list()
    }
    for (ialpha in 1:length(alphas)) {
        alpha <- alphas[ialpha]
        for (f in 1:nfields) {
            field <- mfields[f]
            if (print_comp_kernels) {
                cat(sprintf("computing kernel_%s for alpha=%g\n",
                  field, alpha))
                flush.console()
            }
            if (field == "ind") {
                Km <- 0
                for (type in syb_types) {
                  if (print_comp_kernels)
                    cat(type)
                  Km <- Km + cmf_indicator_kernel_matrix_pred(mdb_pred,
                    mdb_train, alpha, type, verbose = print_comp_kernels)
                }
            }
            else {
                Km <- cmf_kernel_matrix_tp(field, mdb_pred, mdb_train,
                  alpha, verbose = print_comp_kernels)
            }
            kernels_pred[[field]][[ialpha]] <- Km
        }
    }
    save(kernels_pred, file = kernels_pred_fname)
  }

#> function (train_fname = "ligands-train.mol2", pred_fname = "ligands-pred.mol2", 
#>     kernels_pred_fname = "ligands-kernels-pred.RData", mfields = c("q", 
#>         "vdw", "logp", "abra", "abrb"), print_comp_kernels = TRUE, 
#>     ...) 
#> {
#>     mdb0_train <- read_mol2(train_fname)
#>     mdb0_pred <- read_mol2(pred_fname)
#>     mdb_train <- cmf_params_tripos(mdb0_train)
#>     mdb_pred <- cmf_params_tripos(mdb0_pred)
#>     nfields <- length(mfields)
#>     syb_types <- get_syb_types_list(mdb_train)
#>     kernels_pred <- list()
#>     kernels_pred$alphas <- alphas
#>     for (f in 1:nfields) {
#>         kernels_pred[[mfields[f]]] <- list()
#>     }
#>     for (ialpha in 1:length(alphas)) {
#>         alpha <- alphas[ialpha]
#>         for (f in 1:nfields) {
#>             field <- mfields[f]
#>             if (print_comp_kernels) {
#>                 cat(sprintf("computing kernel_%s for alpha=%g\n", 
#>                   field, alpha))
#>                 flush.console()
#>             }
#>             if (field == "ind") {
#>                 Km <- 0
#>                 for (type in syb_types) {
#>                   if (print_comp_kernels) 
#>                     cat(type)
#>                   Km <- Km + cmf_indicator_kernel_matrix_pred(mdb_pred, 
#>                     mdb_train, alpha, type, verbose = print_comp_kernels)
#>                 }
#>             }
#>             else {
#>                 Km <- cmf_kernel_matrix_tp(field, mdb_pred, mdb_train, 
#>                   alpha, verbose = print_comp_kernels)
#>             }
#>             kernels_pred[[field]][[ialpha]] <- Km
#>         }
#>     }
#>     save(kernels_pred, file = kernels_pred_fname)
#>   }
#> <environment: 0x1129ab600>