Computes CMF kernel matrices for training and saves to file

Computes CMF kernel matrices for training and saves to file

comp_kernels_train(train_fname = "ligands-train.mol2", kernels_train_fname = "ligands-kernels-train.RData", mfields = c("q", "vdw", "logp", "abra", "abrb"), print_comp_kernels = TRUE, ...)

Arguments

train_fname

kernels_train_fname

mfields

print_comp_kernels

…

Details

Value

References

Note

See also

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (train_fname = "ligands-train.mol2", kernels_train_fname = "ligands-kernels-train.RData",
    mfields = c("q", "vdw", "logp", "abra", "abrb"), print_comp_kernels = TRUE,
    ...)
{
    mdb0 <- read_mol2(train_fname)
    mdb <- cmf_params_tripos(mdb0)
    nfields <- length(mfields)
    syb_types <- get_syb_types_list(mdb)
    kernels <- list()
    kernels$alphas <- alphas
    for (f in 1:nfields) {
        kernels[[mfields[f]]] <- list()
    }
    for (ialpha in 1:length(alphas)) {
        alpha <- alphas[ialpha]
        for (f in 1:nfields) {
            field <- mfields[f]
            if (print_comp_kernels) {
                cat(sprintf("computing kernel_%s for alpha=%g\n",
                  field, alpha))
                flush.console()
            }
            if (field == "ind") {
                Km <- 0
                for (type in syb_types) {
                  if (print_comp_kernels)
                    cat(type)
                  Km <- Km + cmf_indicator_kernel_matrix(mdb,
                    alpha, type, verbose = print_comp_kernels)
                }
            }
            else {
                Km <- cmf_kernel_matrix_tt(field, mdb, alpha,
                  verbose = print_comp_kernels)
            }
            kernels[[field]][[ialpha]] <- Km
        }
    }
    save(kernels, file = kernels_train_fname)
  }

#> function (train_fname = "ligands-train.mol2", kernels_train_fname = "ligands-kernels-train.RData", 
#>     mfields = c("q", "vdw", "logp", "abra", "abrb"), print_comp_kernels = TRUE, 
#>     ...) 
#> {
#>     mdb0 <- read_mol2(train_fname)
#>     mdb <- cmf_params_tripos(mdb0)
#>     nfields <- length(mfields)
#>     syb_types <- get_syb_types_list(mdb)
#>     kernels <- list()
#>     kernels$alphas <- alphas
#>     for (f in 1:nfields) {
#>         kernels[[mfields[f]]] <- list()
#>     }
#>     for (ialpha in 1:length(alphas)) {
#>         alpha <- alphas[ialpha]
#>         for (f in 1:nfields) {
#>             field <- mfields[f]
#>             if (print_comp_kernels) {
#>                 cat(sprintf("computing kernel_%s for alpha=%g\n", 
#>                   field, alpha))
#>                 flush.console()
#>             }
#>             if (field == "ind") {
#>                 Km <- 0
#>                 for (type in syb_types) {
#>                   if (print_comp_kernels) 
#>                     cat(type)
#>                   Km <- Km + cmf_indicator_kernel_matrix(mdb, 
#>                     alpha, type, verbose = print_comp_kernels)
#>                 }
#>             }
#>             else {
#>                 Km <- cmf_kernel_matrix_tt(field, mdb, alpha, 
#>                   verbose = print_comp_kernels)
#>             }
#>             kernels[[field]][[ialpha]] <- Km
#>         }
#>     }
#>     save(kernels, file = kernels_train_fname)
#>   }
#> <environment: 0x10735ae28>